requirements.txt
requests
beautifulsoup4
transformers
torch
playwright
playwright-stealth
lxml
python-dotenv
sentence-transformers
scikit-learn
hdbscan
sentencepiece
bertopic